pyproject.toml
[tool.poetry]
name = "ecommerce-research-agent"
version = "0.1.0"
description = "AI-powered e-commerce research and market analysis agent"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.13"
pydantic = ">=2.11.9,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
loguru = ">=0.7.3,<0.8.0"
tqdm = ">=4.67.1,<5.0.0"
pandas = ">=2.0.0,<3.0.0"
numpy = ">=1.24.0,<2.0.0"
snowflake-snowpark-python = "^1.39.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.13.1,<0.14.0"
pytest = ">=8.4.2,<9.0.0"

[tool.poetry.scripts]
start-api = "src.api.main:start_server"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 132
indent-width = 4
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
ignore = [
    "E501", # line-too-long - allow long strings, ruff format already wraps statements.
]